jrenteria_Hour1and2
Had to rename but in this project I created a new scene, after deleting the sample scene, called TryItYourself_Hour2. I followed the instructions in the textbook. I also experimented and created a blueprint of my apartment. Then I created a new scene called jrenteria_Hour1and2. I followed the instructions in the module Hour 1 and 2. I created a rough car for this scene.